backend/requirements.txt
psycopg[binary]==3.3.2
requests==2.32.3
python-dotenv==1.0.1